MM-46373 : Move "UserActivityPost" and other types from MM-redux
