+++ components/post_view/post/post.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import {Posts} from 'mattermost-redux/constants';
 import {isMeMessage as checkIsMeMessage} from 'mattermost-redux/utils/post_utils';
-import {UserActivityPost} from 'mattermost-redux/types/posts';
+import {UserActivityPost} from 'mattermost-redux/selectors/entities/posts';
 
 import {makeIsEligibleForClick} from 'utils/utils';
 import * as PostUtils from 'utils/post_utils';
