+++ packages/mattermost-redux/src/selectors/entities/posts.ts
@@ -9,8 +9,6 @@ import {getCurrentUser} from 'mattermost-redux/selectors/entities/common';
 import {getMyPreferences} from 'mattermost-redux/selectors/entities/preferences';
 import {getUsers, getCurrentUserId, getUserStatuses} from 'mattermost-redux/selectors/entities/users';
 
-import {PostWithFormatData} from 'mattermost-redux/types/posts';
-
 import {Channel} from '@mattermost/types/channels';
 import {
     MessageHistory,
@@ -40,6 +38,11 @@ import {getPreferenceKey} from 'mattermost-redux/utils/preference_utils';
 
 export function getAllPosts(state: GlobalState) {
     return state.entities.posts.posts;
+}
+
+export type UserActivityPost = Post & {
+    system_post_ids: string[];
+    user_activity_posts: Post[];
 }
 
 export function getPost(state: GlobalState, postId: Post['id']): Post {
@@ -86,6 +89,17 @@ export function getOpenGraphMetadataForUrl(state: GlobalState, postId: string, u
 export function getPostIdsInCurrentChannel(state: GlobalState): Array<Post['id']> | undefined | null {
     return getPostIdsInChannel(state, state.entities.channels.currentChannelId);
 }
+
+export type PostWithFormatData = Post & {
+    isFirstReply: boolean;
+    isLastReply: boolean;
+    previousPostIsComment: boolean;
+    commentedOnPost?: Post;
+    consecutivePostByUser: boolean;
+    replyCount: number;
+    isCommentMention: boolean;
+    highlight: boolean;
+};
 
 // getPostsInCurrentChannel returns the posts loaded at the bottom of the channel. It does not include older posts
 // such as those loaded by viewing a thread or a permalink.
